// src/lib/generated/portfolio-manifest.ts
export type PortfolioManifestItem = {
  filename: string;
  width: number;
  height: number;
  previewSrc: string;
};